+++ src/js_email_password_auth.hpp
@@ -27,7 +27,7 @@ namespace realm {
 namespace js {
 
 template<typename T>
-class EmailPasswordProviderClientClass : public ClassDefinition<T, app::App::UsernamePasswordProviderClient> {
+class EmailPasswordAuthClass : public ClassDefinition<T, app::App::UsernamePasswordProviderClient> {
     using GlobalContextType = typename T::GlobalContext;
     using ContextType = typename T::Context;
     using FunctionType = typename T::Function;
@@ -41,7 +41,7 @@ class EmailPasswordProviderClientClass : public ClassDefinition<T, app::App::Use
     using Arguments = js::Arguments<T>;
 
 public:
-    std::string const name = "EmailPasswordProviderClient";
+    std::string const name = "EmailPasswordAuth";
 
     static FunctionType create_constructor(ContextType);
     static ObjectType create_instance(ContextType, SharedApp);
@@ -68,21 +68,21 @@ public:
 };
 
 template<typename T>
-inline typename T::Function EmailPasswordProviderClientClass<T>::create_constructor(ContextType ctx) {
-    FunctionType constructor = ObjectWrap<T, EmailPasswordProviderClientClass<T>>::create_constructor(ctx);
+inline typename T::Function EmailPasswordAuthClass<T>::create_constructor(ContextType ctx) {
+    FunctionType constructor = ObjectWrap<T, EmailPasswordAuthClass<T>>::create_constructor(ctx);
     return constructor;
 }
 
 template<typename T>
-typename T::Object EmailPasswordProviderClientClass<T>::create_instance(ContextType ctx, SharedApp app) {
-    return create_object<T, EmailPasswordProviderClientClass<T>>(ctx, new app::App::UsernamePasswordProviderClient(app->provider_client<realm::app::App::UsernamePasswordProviderClient>()));
+typename T::Object EmailPasswordAuthClass<T>::create_instance(ContextType ctx, SharedApp app) {
+    return create_object<T, EmailPasswordAuthClass<T>>(ctx, new app::App::UsernamePasswordProviderClient(app->provider_client<realm::app::App::UsernamePasswordProviderClient>()));
 }
 
 template<typename T>
-void EmailPasswordProviderClientClass<T>::register_user(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
+void EmailPasswordAuthClass<T>::register_user(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
     args.validate_count(3);
 
-    auto& client = *get_internal<T, EmailPasswordProviderClientClass<T>>(ctx, this_object);
+    auto& client = *get_internal<T, EmailPasswordAuthClass<T>>(ctx, this_object);
 
     auto email = Value::validated_to_string(ctx, args[0], "email");
     auto password = Value::validated_to_string(ctx, args[1], "password");
@@ -92,10 +92,10 @@ void EmailPasswordProviderClientClass<T>::register_user(ContextType ctx, ObjectT
 }
 
 template<typename T>
-void EmailPasswordProviderClientClass<T>::confirm_user(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
+void EmailPasswordAuthClass<T>::confirm_user(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
     args.validate_count(3);
 
-    auto& client = *get_internal<T, EmailPasswordProviderClientClass<T>>(ctx, this_object);
+    auto& client = *get_internal<T, EmailPasswordAuthClass<T>>(ctx, this_object);
 
     auto token = Value::validated_to_string(ctx, args[0], "token");
     auto token_id = Value::validated_to_string(ctx, args[1], "token_id");
@@ -105,10 +105,10 @@ void EmailPasswordProviderClientClass<T>::confirm_user(ContextType ctx, ObjectTy
 }
 
 template<typename T>
-void EmailPasswordProviderClientClass<T>::resend_confirmation_email(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
+void EmailPasswordAuthClass<T>::resend_confirmation_email(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
     args.validate_count(2);
 
-    auto& client = *get_internal<T, EmailPasswordProviderClientClass<T>>(ctx, this_object);
+    auto& client = *get_internal<T, EmailPasswordAuthClass<T>>(ctx, this_object);
 
     auto email = Value::validated_to_string(ctx, args[0], "email");
     auto callback = Value::validated_to_function(ctx, args[1], "callback");
@@ -117,10 +117,10 @@ void EmailPasswordProviderClientClass<T>::resend_confirmation_email(ContextType 
 }
 
 template<typename T>
-void EmailPasswordProviderClientClass<T>::send_reset_password_email(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
+void EmailPasswordAuthClass<T>::send_reset_password_email(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
     args.validate_count(2);
 
-    auto& client = *get_internal<T, EmailPasswordProviderClientClass<T>>(ctx, this_object);
+    auto& client = *get_internal<T, EmailPasswordAuthClass<T>>(ctx, this_object);
 
     auto email = Value::validated_to_string(ctx, args[0], "email");
     auto callback = Value::validated_to_function(ctx, args[1], "callback");
@@ -129,10 +129,10 @@ void EmailPasswordProviderClientClass<T>::send_reset_password_email(ContextType 
 }
 
 template<typename T>
-void EmailPasswordProviderClientClass<T>::reset_password(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
+void EmailPasswordAuthClass<T>::reset_password(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
     args.validate_count(4);
 
-    auto& client = *get_internal<T, EmailPasswordProviderClientClass<T>>(ctx, this_object);
+    auto& client = *get_internal<T, EmailPasswordAuthClass<T>>(ctx, this_object);
 
     auto password = Value::validated_to_string(ctx, args[0], "password");
     auto token = Value::validated_to_string(ctx, args[1], "token");
@@ -143,10 +143,10 @@ void EmailPasswordProviderClientClass<T>::reset_password(ContextType ctx, Object
 }
 
 template<typename T>
-void EmailPasswordProviderClientClass<T>::call_reset_password_function(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
+void EmailPasswordAuthClass<T>::call_reset_password_function(ContextType ctx, ObjectType this_object, Arguments& args, ReturnValue& return_value) {
     args.validate_count(4);
 
-    auto& client = *get_internal<T, EmailPasswordProviderClientClass<T>>(ctx, this_object);
+    auto& client = *get_internal<T, EmailPasswordAuthClass<T>>(ctx, this_object);
 
     auto email = Value::validated_to_string(ctx, args[0], "email");
     auto password = Value::validated_to_string(ctx, args[1], "password");