+++ tests/js/user-tests.js
@@ -92,7 +92,7 @@ function randomNonVerifiableEmail() {
 async function registerAndLogInEmailUser(app) {
   const validEmail = randomVerifiableEmail();
   const validPassword = "test1234567890";
-  await app.auth.emailPassword.registerUser(validEmail, validPassword);
+  await app.emailPasswordAuth.registerUser(validEmail, validPassword);
   let user = await app.logIn(Realm.Credentials.emailPassword(validEmail, validPassword))
   assertIsUser(user);
   assertIsSameUser(user, app.currentUser);
@@ -141,10 +141,10 @@ module.exports = {
     });
   },
 
-  async testEmailPasswordProvider() {
+  async testEmailPasswordAuth() {
     let app = new Realm.App(appConfig);
-    let provider = app.auth.emailPassword;
-    TestCase.assertTrue(provider instanceof Realm.Auth.EmailPasswordProvider);
+    let provider = app.emailPasswordAuth;
+    TestCase.assertTrue(provider instanceof Realm.Auth.EmailPasswordAuth);
   },
 
   async testRegisterAutoVerifyEmailPassword() {
@@ -158,7 +158,7 @@ module.exports = {
       let credentials = Realm.Credentials.emailPassword(invalidEmail, invalidPassword);
       let err = await TestCase.assertThrowsAsync(async() => app.logIn(credentials));
       TestCase.assertEqual(err.message, "invalid username/password"); // this user does not exist yet
-      err = await TestCase.assertThrowsAsync(async() => app.auth.emailPassword.registerUser(invalidEmail, invalidPassword));
+      err = await TestCase.assertThrowsAsync(async() => app.emailPasswordAuth.registerUser(invalidEmail, invalidPassword));
       TestCase.assertEqual(err.message, "password must be between 6 and 128 characters");
       err = await TestCase.assertThrowsAsync(async() => app.logIn(credentials));
       TestCase.assertEqual(err.message, "invalid username/password"); // this user did not register
@@ -167,7 +167,7 @@ module.exports = {
       let credentials = Realm.Credentials.emailPassword(invalidEmail, validPassword);
       let err = await TestCase.assertThrowsAsync(async() => app.logIn(credentials));
       TestCase.assertEqual(err.message, "invalid username/password"); // this user does not exist yet
-      err = await TestCase.assertThrowsAsync(async() => app.auth.emailPassword.registerUser(invalidEmail, validPassword));
+      err = await TestCase.assertThrowsAsync(async() => app.emailPasswordAuth.registerUser(invalidEmail, validPassword));
       TestCase.assertEqual(err.message, `failed to confirm user ${invalidEmail}`);
       err = await TestCase.assertThrowsAsync(async() => app.logIn(credentials));
       TestCase.assertEqual(err.message, "invalid username/password"); // this user did not register
@@ -176,7 +176,7 @@ module.exports = {
       let credentials = Realm.Credentials.emailPassword(validEmail, invalidPassword);
       let err = await TestCase.assertThrowsAsync(async() => app.logIn(credentials));
       TestCase.assertEqual(err.message, "invalid username/password"); // this user does not exist yet
-      err = await TestCase.assertThrowsAsync(async() => app.auth.emailPassword.registerUser(validEmail, invalidPassword));
+      err = await TestCase.assertThrowsAsync(async() => app.emailPasswordAuth.registerUser(validEmail, invalidPassword));
       TestCase.assertEqual(err.message, "password must be between 6 and 128 characters");
       err = await TestCase.assertThrowsAsync(async() => app.logIn(credentials));
       TestCase.assertEqual(err.message, "invalid username/password"); // this user did not register
@@ -185,7 +185,7 @@ module.exports = {
       let credentials = Realm.Credentials.emailPassword(validEmail, validPassword);
       let err = await TestCase.assertThrowsAsync(async() => app.logIn(credentials));
       TestCase.assertEqual(err.message, "invalid username/password"); // this user does not exist yet
-      await app.auth.emailPassword.registerUser(validEmail, validPassword);
+      await app.emailPasswordAuth.registerUser(validEmail, validPassword);
       let user = await app.logIn(credentials)
       assertIsUser(user);
       assertIsSameUser(user, app.currentUser);