+++ CHANGELOG.md
@@ -12,6 +12,8 @@ NOTE: This version bumps the Realm file format to version 11. It is not possible
 * `Realm.Auth.UserAPIKeyProvider` has been replaced by `Realm.Auth.ApiKeyProvider`.
 * `user.auth.apiKeys` has been replaced by `user.apiKeys`.
 *  The instance methods on the ApiKeyAuth instance (`user.apiKeys`) have gotten their APIKey(s) suffix removed: Ex. `apiKeys.createAPIKey` has been replaced by `apiKeys.create`.
+* `Realm.Auth.EmailPasswordProvider` has been replaced by `Realm.Auth.EmailPasswordAuth`.
+* `app.auth.emailPassword` has been replaced by `user.emailPasswordAuth`.
 
 ### Enhancements
 * Added RemoteMongoClient functionality to `Realm.User`